fix(desktop): stop one unsupported respond-to mode hiding every agent
The directory projection accepted "nobody" and forwarded it verbatim.
RelayAgentInfo deserializes respond_to into RespondTo, which deliberately has
no nobody variant because the desktop exposes no surface for it, so a single
validly signed event carrying that mode failed the whole Vec and every other
agent vanished from discovery.

Project only the modes RespondTo can represent. An unrepresentable mode becomes
absent, and every eligibility check treats absent as not-invocable, so the
projection degrades closed rather than either crashing or over-granting.
