+++ desktop/src-tauri/src/nostr_convert.rs
@@ -22,6 +22,9 @@ pub use user_search::{
 
 // ── Tag helpers ─────────────────────────────────────────────────────────────
 
+mod managed_agents;
+pub use managed_agents::managed_agents_from_events;
+
 /// Find the first tag whose name matches `name` and return its first value.
 ///
 /// e.g. for tag `["name", "general"]` with `name="name"` returns `Some("general")`.
@@ -495,69 +498,6 @@ pub fn agents_from_events(events: &[Event]) -> Value {
     json!({ "agents": arr })
 }
 
-// ── kind:30177 (managed-agent directory) ───────────────────────────────────
-
-/// Convert public managed-agent projections into the relay-agent directory
-/// shape consumed by the desktop.
-///
-/// The event author is the human owner. The managed agent's public key is the
-/// parameterized replaceable event's `d` tag, so a content-supplied key is
-/// never authoritative. Build a narrow output object instead of forwarding
-/// the complete content projection so future fields cannot accidentally cross
-/// this frontend boundary.
-pub fn managed_agents_from_events(events: &[Event]) -> Value {
-    let arr: Vec<Value> = events
-        .iter()
-        .filter_map(|event| {
-            let agent_pubkey = nostr::PublicKey::from_hex(first_tag_value(event, "d")?).ok()?;
-            let pubkey = agent_pubkey.to_hex();
-            let npub = agent_pubkey.to_bech32().unwrap_or_else(|_| pubkey.clone());
-            let content: Value = serde_json::from_str(&event.content).ok()?;
-            let object = content.as_object()?;
-            let name = object
-                .get("name")
-                .and_then(Value::as_str)
-                .map(str::trim)
-                .filter(|value| !value.is_empty())
-                .unwrap_or(&npub);
-            let respond_to = object
-                .get("respond_to")
-                .and_then(Value::as_str)
-                .filter(|mode| matches!(*mode, "owner-only" | "allowlist" | "anyone" | "nobody"));
-
-            let mut respond_to_allowlist = Vec::new();
-            if let Some(values) = object.get("respond_to_allowlist").and_then(Value::as_array) {
-                for value in values {
-                    let Some(raw) = value.as_str() else {
-                        continue;
-                    };
-                    let Ok(key) = nostr::PublicKey::from_hex(raw.trim()) else {
-                        continue;
-                    };
-                    let normalized = key.to_hex();
-                    if !respond_to_allowlist.contains(&normalized) {
-                        respond_to_allowlist.push(normalized);
-                    }
-                }
-            }
-
-            Some(json!({
-                "pubkey": pubkey,
-                "name": name,
-                "agent_type": "agent",
-                "channels": [],
-                "channel_ids": [],
-                "capabilities": [],
-                "status": "offline",
-                "respond_to": respond_to,
-                "respond_to_allowlist": respond_to_allowlist,
-            }))
-        })
-        .collect();
-
-    json!({ "agents": arr })
-}
-
 // ── kind:13534 (relay membership list) ──────────────────────────────────────
 
 /// Convert a kind:13534 relay membership list to the relay members format.
@@ -1024,53 +964,6 @@ mod tests {
             Some(crate::managed_agents::RespondTo::Allowlist)
         );
         assert_eq!(parsed[0].respond_to_allowlist, vec!["a".repeat(64)]);
-    }
-
-    #[test]
-    fn managed_agents_use_d_tag_identity_and_preserve_allowlist_metadata() {
-        let agent_pubkey = "02".repeat(32);
-        let allowed_pubkey = "03".repeat(32);
-        let e = ev(
-            30177,
-            &format!(
-                r#"{{"pubkey":"forged","name":"Scout","parallelism":1,"respond_to":"allowlist","respond_to_allowlist":["{allowed_pubkey}"],"env_vars":{{"SECRET":"do-not-project"}}}}"#
-            ),
-            vec![vec!["d", &agent_pubkey]],
-        );
-        let v = managed_agents_from_events(std::slice::from_ref(&e));
-        let agents = v.get("agents").cloned().unwrap();
-        let parsed: Vec<crate::managed_agents::RelayAgentInfo> =
-            serde_json::from_value(agents).unwrap();
-
-        assert_eq!(parsed.len(), 1);
-        assert_eq!(parsed[0].pubkey, agent_pubkey);
-        assert_eq!(parsed[0].name, "Scout");
-        assert_eq!(
-            parsed[0].respond_to,
-            Some(crate::managed_agents::RespondTo::Allowlist)
-        );
-        assert_eq!(parsed[0].respond_to_allowlist, vec![allowed_pubkey]);
-        assert!(
-            !v.to_string().contains("SECRET"),
-            "the directory projection must remain an explicit public-field allowlist"
-        );
-    }
-
-    #[test]
-    fn managed_agents_drop_events_without_a_valid_agent_d_tag() {
-        let missing = ev(
-            30177,
-            r#"{"name":"Missing","parallelism":1,"respond_to":"owner-only"}"#,
-            vec![],
-        );
-        let invalid = ev(
-            30177,
-            r#"{"name":"Invalid","parallelism":1,"respond_to":"owner-only"}"#,
-            vec![vec!["d", "not-a-pubkey"]],
-        );
-        let v = managed_agents_from_events(&[missing, invalid]);
-
-        assert_eq!(v.get("agents").and_then(Value::as_array).unwrap().len(), 0);
     }
 
     #[test]
